Cleaner apiCalls
Moved the api call implementations into ApiService so the coe is more moular and organized.
